requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
abseil-cpp=20210324.2=hd77b12b_0
absl-py=0.13.0=py38haa95532_0
aiohttp=3.7.4.post0=py38h2bbff1b_2
astor=0.8.1=py38haa95532_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38haa95532_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38haa95532_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.10.8=h5b45459_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py38haa244fe_1
cffi=1.15.0=py38h2bbff1b_0
chardet=4.0.0=py38haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.8=py38h71e12ea_0
cudatoolkit=11.3.1=h59b6b97_2
cudnn=8.2.1=cuda11.3_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.0=pyhd3eb1b0_0
entrypoints=0.3=py38_0
flatbuffers=2.0.0=h6c2663c_0
fonttools=4.28.2=pypi_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h62dcd97_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py38hc60d5dd_0
h5py=3.6.0=py38h3de5c98_0
hdf5=1.10.6=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.1=h6c2663c_0
idna=3.2=pyhd3eb1b0_0
importlib-metadata=4.8.1=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.4.1=py38haa95532_1
ipython=7.29.0=py38hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py38haa95532_1
joblib=1.1.0=pypi_0
jpeg=9d=h2bbff1b_0
jupyter_client=7.0.6=pyhd3eb1b0_0
jupyter_core=4.9.1=py38haa95532_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.2=pypi_0
libcurl=7.78.0=h86230a5_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libssh2=1.9.0=h7a1dbc1_1
markdown=3.3.4=py38haa95532_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
multidict=5.1.0=py38h2bbff1b_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
numpy=1.21.2=py38hfca59bb_0
numpy-base=1.21.2=py38h0829f74_0
oauthlib=3.1.1=pyhd3eb1b0_0
opencv-python=4.5.4.60=pypi_0
openssl=1.1.1l=h8ffe710_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pypi_0
parso=0.8.2=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=pypi_0
pip=21.2.2=py38haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.14.0=py38hd77b12b_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.10.0=pyhd3eb1b0_0
pyjwt=2.1.0=py38haa95532_0
pymaxflow=1.2.13=py38h5d928e2_1
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.6=pypi_0
pyreadline=2.1=py38_1
pysocks=1.7.1=py38haa95532_0
python=3.8.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pywin32=228=py38hbaba5e8_1
pyzmq=22.2.1=py38hd77b12b_1
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.1=pypi_0
scipy=1.7.1=py38hbe87c03_2
setuptools=58.0.4=py38haa95532_0
setuptools-scm=6.3.2=pypi_0
six=1.16.0=pyhd3eb1b0_0
snappy=1.1.8=h33f27b4_0
spyder-kernels=1.9.2=py38_0
sqlite=3.36.0=h2bbff1b_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.5.0=gpu_py38h8e8c102_0
tensorflow-base=2.5.0=gpu_py38hb3da07e_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
tensorflow-gpu=2.5.0=h17022bd_0
termcolor=1.1.0=py38haa95532_1
threadpoolctl=3.0.0=pypi_0
tomli=1.2.2=pypi_0
tornado=6.1=py38h2bbff1b_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.35.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
wrapt=1.12.1=py38he774522_1
yarl=1.6.3=py38h2bbff1b_0
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
